feat: implement resilient ID-anchored tooltip resolution and expand card indexer

+++ engine_rust_src/src/core/logic/interpreter/logging.rs
@@ -1,7 +1,7 @@
 use crate::core::generated_constants::*;
 use crate::core::enums::*;
 
-pub fn get_opcode_log(op: i32, v: i32, a: i32, _s: i32, result_count: i32) -> Option<String> {
+pub fn get_opcode_log(op: i32, v: i32, a: i64, _s: i32, result_count: i32) -> Option<String> {
     match op {
         O_DRAW => Some(format!("Draw {} card(s)", v)),
         O_ADD_HEARTS => {
